backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.2
langchain==0.2.1
langchain-google-genai==1.0.5
python-dotenv==1.0.1
chromadb==0.5.0
sentence-transformers==2.7.0
